Fixed 0020024: Error message 'Data saving problem' when try to checkout with multiple addresses

+++ app/code/core/Mage/Checkout/Model/Type/Multishipping.php
@@ -226,7 +226,7 @@ class Mage_Checkout_Model_Type_Multishipping extends Mage_Checkout_Model_Type_Ab
              * MultisippingQty should be defined for each quote item when it processed with _addShippingItem
              */
             foreach ($quote->getAllItems() as $_item) {
-                if (!$_item->getProduct()->getIsVirtual() && !$_item->getMultisippingQty()) {
+                if (!$_item->getProduct()->getIsVirtual() && !$_item->getParentItem() && !$_item->getMultisippingQty()) {
                     $_item->delete();
                 }
             }
